Let OpenGl be accessed through a server
The Renderer takes a GlServer server instance and uses it for rendering.
With this approach a user implements a GlServer on top of the graphics library
they want to use and pass that server to the renderer. The platform folder
implements a WebGl server for now that can be used as a GlServer.

For Linyard to be truly platform neutral its components should delegate platform
dependent responsibilities to the user. Removing the Environment enum from renderer
is part of this delegation

Other changes:
>Add added a stylesheet for web apps